src/stream_analysis/pyproject.toml
[tool.poetry]
name = "stream-analysis"
version = "0.1.0"
description = ""
authors = ["DaiichiroNoda <69710300+donachii@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ffmpeg-python = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
